requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
six==1.16.0
spotipy==2.22.0
urllib3==1.26.14
